added few developers' tools
git-svn-id: https://quantlib.svn.sourceforge.net/svnroot/quantlib/trunk/QuantLib@1628 8618b1d8-e22c-0410-b026-96a5dba3e089

+++ ql/Math/segmentintegral.hpp
@@ -50,11 +50,11 @@ namespace QuantLib {
         */
         class SegmentIntegral{
           public:
-            SegmentIntegral(long intervals);
+            SegmentIntegral(size_t intervals);
             double operator()(const ObjectiveFunction &f, 
                              double a, double b) const;
           private:
-            long intervals_;
+            size_t intervals_;
        };
 
     }
